requirements.txt
asgiref==3.8.1
Django==5.2.1
pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
weasyprint==66.0
gunicorn==21.2.0
python-dotenv==1.0.1
